requirements.txt
Django==3.1.12
djangorestframework==3.12.4
django-cors-headers==3.7.0
django-extensions==3.1.3
drf-yasg==1.20.0
django-storages==1.11.1
djongo==1.3.6
dnspython==2.1.0
python-dotenv==0.19.0
pyjwt==2.1.0
requests==2.26.0
pymongo==3.7.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
google-cloud-storage==1.41.1
whitenoise==5.3.0
gunicorn==20.1.0
